% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df_index.R
\name{make_df_index}
\alias{make_df_index}
\title{Utility function to map table with their columns/rows in a bigger table}
\usage{
make_df_index(v)
}
\arguments{
\item{v}{import Actigraphy Raw data}
}
\value{
The lengths and values of runs of equal values in a vector.
}
\description{
Utility function to map table with their columns/rows in a bigger table
}
