% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df.R
\name{split_df}
\alias{split_df}
\title{Split a large table into smaller tables if there are blank columns or rows (if you still see entire rows or columns missing. Please increase complexity)}
\usage{
split_df(df, showWarnig = TRUE, complexity = 1)
}
\arguments{
\item{df}{Actigraphy Raw data}
}
\value{
The lengths and values of runs of equal values in a vector.
}
\description{
Split a large table into smaller tables if there are blank columns or rows (if you still see entire rows or columns missing. Please increase complexity)
}
