% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_direction.R
\name{split_direction}
\alias{split_direction}
\title{Split a large table in one direction if there are blank columns or rows}
\usage{
split_direction(df, direction = "col")
}
\arguments{
\item{df}{Actigraphy Raw data}
}
\value{
The lengths and values of runs of equal values in a vector.
}
\description{
Split a large table in one direction if there are blank columns or rows
}
