% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_rle.R
\name{vec_rle}
\alias{vec_rle}
\title{Utility function to get rle as a named vector}
\usage{
vec_rle(v)
}
\arguments{
\item{v}{import Actigraphy Raw data}
}
\value{
The lengths and values of runs of equal values in a vector.
}
\description{
Utility function to get rle as a named vector
}
