% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractdata.R
\name{extractdata}
\alias{extractdata}
\title{Extract Actigraphy Data to csv file}
\usage{
extractdata(table_raw, initalize)
}
\arguments{
\item{table_raw}{import Actigraphy Raw data}

\item{initalize}{import Initalize Raw data}
}
\value{
The final result.
}
\description{
Extract Actigraphy Data to csv file
}
